+++ .dotnet/api/OpenAI.netstandard2.0.cs
@@ -1661,12 +1661,12 @@ namespace OpenAI.Chat {
 namespace OpenAI.Embeddings {
     public class Embedding : IJsonModel<Embedding>, IPersistableModel<Embedding> {
         public int Index { get; }
-        public ReadOnlyMemory<float> Vector { get; }
         Embedding IJsonModel<Embedding>.Create(ref Utf8JsonReader reader, ModelReaderWriterOptions options);
         void IJsonModel<Embedding>.Write(Utf8JsonWriter writer, ModelReaderWriterOptions options);
         Embedding IPersistableModel<Embedding>.Create(BinaryData data, ModelReaderWriterOptions options);
         string IPersistableModel<Embedding>.GetFormatFromOptions(ModelReaderWriterOptions options);
         BinaryData IPersistableModel<Embedding>.Write(ModelReaderWriterOptions options);
+        public ReadOnlyMemory<float> ToFloats();
     }
     public class EmbeddingClient {
         protected EmbeddingClient();